Update excalidraw page components. We can keep these exposed in dev mode (:3000) for testing.
